- Move business logic for generating rad data file to a dao module class - Configure cronjob - Move bucket names to settings.py vars
