+++ data_aggregator/dao.py
@@ -2,8 +2,8 @@ import logging
 import os
 from django.conf import settings
 from data_aggregator.models import Assignment, Course, Participation, \
-    Term, User, Week
-from data_aggregator.utilities import get_week_of_term
+    Term, User, Week, RadDbView
+from data_aggregator.utilities import get_week_of_term, get_view_name
 from restclients_core.exceptions import DataFailureException
 from restclients_core.util.retry import retry
 from uw_sws.term import get_current_term
@@ -13,6 +13,14 @@ from uw_canvas.courses import Courses
 from uw_canvas.enrollments import Enrollments
 from uw_canvas.reports import Reports
 from uw_canvas.terms import Terms
+
+import numpy as np
+import pandas as pd
+from io import IOBase, StringIO, BytesIO
+from boto3 import client
+from google.cloud import storage
+from google.cloud.exceptions import NotFound
+from datetime import datetime, timezone
 
 
 class AnalyticTypes():
@@ -304,3 +312,161 @@ class CanvasDAO():
         sis_data = self.reports.get_report_data(user_report)
         self.reports.delete_report(user_report)
         return sis_data
+
+
+class LoadRadDao():
+
+    def __init__(self):
+        self._configure_pandas()
+        self.gcs_client = storage.Client()
+        self.gcs_bucket_name = settings.RAD_METADATA_BUCKET_NAME
+        self.s3_client = client('s3',
+                                aws_access_key_id=settings.AWS_ACCESS_ID,
+                                aws_secret_access_key=settings.AWS_ACCESS_KEY)
+        self.s3_bucket_name = settings.IDP_BUCKET_NAME
+        sws_term = get_current_term()
+        self.curr_term = sws_term.canvas_sis_id()
+        self.curr_week = get_week_of_term(sws_term.first_day_quarter)
+
+    def _configure_pandas(self):
+        pd.options.mode.use_inf_as_na = True
+        pd.options.display.max_rows = 500
+        pd.options.display.precision = 3
+        pd.options.display.float_format = '{:.3f}'.format
+
+    def _zero_range(self, x):
+        '''
+        Check range of series x is non-zero.
+        Helper for rescaling.
+        May choke if all x is na.
+        '''
+        zr = False
+        if x.min() == x.max():
+            zr = True
+        return zr
+
+    def _rescale_range(self, x, min=-5, max=5):
+        '''
+        Scale numbers to an arbitray min/max range.
+        '''
+        if self._zero_range(x):
+            return np.mean([min, max])
+        elif x.isnull().all():
+            return np.nan
+        else:
+            x += -(x.min())
+            x /= x.max() / (max - min)
+            x += min
+            return x
+
+    def download_from_gcs_bucket(self, url_key):
+        bucket = self.gcs_client.get_bucket(self.gcs_bucket_name)
+        try:
+            blob = bucket.get_blob(url_key)
+            content = blob.download_as_string()
+            if content:
+                return content.decode('utf-8')
+        except NotFound as ex:
+            logging.error("gcp {}: {}".format(url_key, ex))
+            raise
+
+    def download_from_s3_bucket(self, url_key):
+        idp_obj = self.s3_client.get_object(Bucket=self.s3_bucket_name,
+                                            Key=url_key)
+        content = BytesIO(idp_obj['Body'].read())
+        return content
+
+    def upload_to_gcs_bucket(self, url_key, content):
+        """
+        Upload a string or file-like object contents to GCS bucket
+
+        :param url_key: URL response to cache
+        :type url_key: str
+        :param content: Content to cache
+        :type content: str or file object
+        """
+        bucket = self.gcs_client.get_bucket(self.gcs_bucket_name)
+        blob = bucket.get_blob(url_key)
+        if not blob:
+            blob = bucket.blob(url_key)
+        blob.custom_time = datetime.now(timezone.utc)
+        if isinstance(content, IOBase):
+            blob.upload_from_file(content,
+                                  num_retries=5,
+                                  timeout=30)
+        else:
+            blob.upload_from_string(str(content),
+                                    num_retries=5,
+                                    timeout=30)
+
+    def get_student_categories_df(self):
+        users = User.objects.all().values("canvas_user_id", "login_id")
+        users_df = pd.DataFrame(users)
+
+        url_key = ("application_metadata/student_categories/"
+                   "{}-netid-name-stunum-categories.csv"
+                   .format(self.curr_term))
+        content = self.download_from_gcs_bucket(url_key)
+
+        sdb_df = pd.read_csv(StringIO(content))
+        i = list(sdb_df.columns[sdb_df.columns.str.startswith('regis_')])
+        i.extend(['yrq', 'enroll_status', 'dept_abbrev',
+                 'course_no', 'section_id', 'course_id'])
+        sdb_df.drop(columns=i, inplace=True)
+        sdb_df.drop_duplicates(inplace=True)
+        sdb_df = sdb_df.merge(users_df.rename(
+            columns={'login_id': 'uw_netid'}), how='left', on='uw_netid')
+        sdb_df.fillna(value={'canvas_user_id': -99}, inplace=True)
+        return sdb_df
+
+    def get_pred_proba_scores_df(self):
+        url_key = ("application_metadata/predicted_probabilites/"
+                   "{}-pred-proba.csv".format(self.curr_term))
+        content = self.download_from_gcs_bucket(url_key)
+
+        probs_df = pd.read_csv(
+            StringIO(content),
+            usecols=['system_key', 'pred0'])
+        probs_df['pred0'] = probs_df['pred0'].transform(self._rescale_range)
+        return probs_df
+
+    def get_rad_dbview_df(self):
+        view_name = get_view_name(self.curr_term, self.curr_week, "rad")
+        rad_db_model = RadDbView.setDb_table(view_name)
+        rad_canvas_qs = rad_db_model.objects.all().values()
+        return pd.DataFrame(rad_canvas_qs)
+
+    def get_idp_df(self):
+        '''
+        Returns a pandas dataframe containing the contents of the
+        last idp object found in the s3 bucket.
+        '''
+        bucket_objects = \
+            self.s3_client.list_objects_v2(Bucket=self.s3_bucket_name)
+        last_idp_file = bucket_objects['Contents'][-1]['Key']
+        logging.info(f'Using {last_idp_file} as idp file.')
+        content = self.download_from_s3_bucket(last_idp_file)
+        idp_df = pd.read_csv(content,
+                             header=0,
+                             names=['uw_netid', 'sign_in'])
+        # normalize sign-in score
+        idp_df['sign_in'] = np.log(idp_df['sign_in']+1)
+        idp_df['sign_in'] = self._rescale_range(idp_df['sign_in'])
+        return idp_df
+
+    def get_rad_cavas_df(self):
+        # get rad canvas data
+        rad_canvas_df = self.get_rad_dbview_df()
+        # get student categories
+        sdb_df = self.get_student_categories_df()
+        # get idp data
+        idp_df = self.get_idp_df()
+        # get predicted probabilities
+        probs_df = self.get_pred_proba_scores_df()
+
+        # merge to create the final dataset
+        joined_canvas_df = (pd.merge(sdb_df, rad_canvas_df, how='left',
+                                     on='canvas_user_id')
+                              .merge(idp_df, how='left', on='uw_netid')
+                              .merge(probs_df, how='left', on='system_key'))
+        return joined_canvas_df
