+++ data_aggregator/management/commands/create_rad_data_file.py
@@ -1,0 +1,16 @@
+from django.core.management.base import BaseCommand
+from data_aggregator.dao import LoadRadDao
+
+
+class Command(BaseCommand):
+
+    help = ("Loads normalized Canvas assignment and participation analytics "
+            "into RAD.")
+
+    def handle(self, *args, **options):
+        lrd = LoadRadDao()
+        rcd = lrd.get_rad_cavas_df()
+        file_name = "rad_data/{}-week-{}-rad-data.csv".format(lrd.curr_term,
+                                                              lrd.curr_week)
+        file_obj = rcd.to_csv(sep=",", index=False, encoding="UTF-8")
+        lrd.upload_to_gcs_bucket(file_name, file_obj)
